requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via starlette
bcrypt==5.0.0
    # via passlib
click==8.3.2
    # via uvicorn
ecdsa==0.19.2
    # via python-jose
fastapi==0.135.3
    # via -r requirements.in
greenlet==3.3.2
    # via sqlalchemy
h11==0.16.0
    # via uvicorn
idna==3.11
    # via anyio
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pyasn1==0.6.3
    # via
    #   python-jose
    #   rsa
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.41.5
    # via pydantic
python-dotenv==1.2.2
    # via -r requirements.in
python-jose==3.5.0
    # via -r requirements.in
python-multipart==0.0.22
    # via -r requirements.in
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via ecdsa
sqlalchemy==2.0.49
    # via -r requirements.in
starlette==1.0.0
    # via fastapi
typing-extensions==4.15.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
uvicorn==0.43.0
    # via -r requirements.in
